requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.0.1
google-api-python-client==2.24.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
O365==2.0.16
oauthlib==3.1.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
stringcase==1.2.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.7
